# swagger.yaml
openapi: 3.0.0
info:
  title: WhSmith
  contact: {}
  version: '1.0'
servers:
- url: https://avnology.org
  variables: {}
- url: http://localhost:1812
  variables: {}
paths:
...
  /product/add:
    post:
      tags:
      - Product
      summary: Add Product
      operationId: AddProduct
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - name[ar]
              - name[en]
              - description[ar]
              - description[en]
              - quantity
              - stores[0][store]
              - stores[0][price]
              - stores[1][store]
              - stores[1][price]
              - images
              - category
              type: object
              properties:
                name[ar]:
                  type: string
                  example: 'تست '
                name[en]:
                  type: string
                  example: 'Test '
                description[ar]:
                  type: string
                  example: تسسيسشيسش سيشسيسشيسشيسشيسشيشسيسشيسشيسش
                description[en]:
                  type: string
                  example: Madddasasda566999ada  asdasdasd5as scadsadas adasdafds
                quantity:
                  type: integer
                  format: int32
                  example: 20
                stores[0][store]:
                  type: string
                  example: 65353008153db4d278762536
                stores[0][price]:
                  type: integer
                  format: int32
                  example: 3000
                stores[1][store]:
                  type: string
                  example: 652579906f0b66edd6aa3c87
                stores[1][price]:
                  type: integer
                  format: int32
                  example: 500
                images:
                  type: string
                  format: binary
                category:
                  type: string
                  example: 653504bbb4c91d4b7cd78469
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /product/one/{id}:
    get:
      tags:
      - Product
      summary: Get Specific Product
      operationId: GetSpecificProduct
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /product/update/{id}:
    patch:
      tags:
      - Product
      summary: Update Product
      operationId: UpdateProduct
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - images
              - stores[0][store]
              - stores[0][price]
              type: object
              properties:
                images:
                  type: string
                  format: binary
                stores[0][store]:
                  type: string
                  example: 652579906f0b66edd6aa3c87
                stores[0][price]:
                  type: integer
                  format: int32
                  example: 900
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /cart/remove/{itemID}:
    delete:
      tags:
      - Cart
      summary: Remove Cart
      operationId: RemoveCart
      parameters:
      - name: itemID
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /order/add/{cartID}:
    post:
      tags:
      - Order
      summary: Add Order
      operationId: AddOrder
      parameters:
      - name: cartID
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /order/dashboard/update/{id}:
    patch:
      tags:
      - Order
      summary: Update Status Order
      operationId: UpdateStatusOrder
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateStatusOrderRequest'
              - example:
                  status: Delivered
            example:
              status: Delivered
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /order/dashboard/{id}:
    get:
      tags:
      - Order
      summary: Get Specific Order
      operationId: GetSpecificOrder
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: string
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    UpdateStatusOrderRequest:
      title: UpdateStatusOrderRequest
      required:
      - status
      type: object
      properties:
        status:
          type: string
      example:
        status: Delivered
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...